+++ Code/src/UnitePhysic.cpp
@@ -49,7 +49,7 @@ b2Body* UnitePhysic::getBox()
 }
 
 //Permet de modifier la position de la boite de collision
-void UnitePhysic::setXY(int _x, int _y)
+void UnitePhysic::setXY(float _x, float _y)
 {
 	Unite::setXY(_x, _y);
     //std::cout << "Unite : (" << Unite::getX() << ", " << Unite::getY() << ")" << endl;
