+++ Code/src/UnitePhysic.h
@@ -11,7 +11,7 @@ public:
 	virtual ~UnitePhysic();		// Le destructeur doit supprimer les boîtes de collisions 
 	UnitePhysic(float _x, float _y, float _h, float _l, std::string textureName, std::map<std::string, sf::Texture>& textDictionnary, b2BodyType bodyType, b2World &world);
 	b2Body* getBox();
-	void setXY(int _x, int _y) override;	// doit également changer les x et y de la boite physique
+	void setXY(float _x, float _y) override;	// doit également changer les x et y de la boite physique
 
 private:
 	b2Body* box;	// Pointeur vers la boite associée à l'unité
